requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==8.1.7
colorama==0.4.6
dj-database-url==2.1.0
dj-rest-auth==6.0.0
Django==5.0.4
django-allauth==0.62.1
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
gunicorn==22.0.0
idna==3.7
joblib==1.4.0
mathparse==0.1.2
nltk==3.8.1
packaging==24.0
Pint==0.23
psycopg2-binary==2.9.9
pymongo==3.13.0
python-dateutil==2.7.5
pytz==2024.1
PyYAML==3.13
regex==2024.4.16
requests==2.31.0
six==1.16.0
SQLAlchemy==1.2.19
sqlparse==0.5.0
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
